refactor, use pushEOSCAction, move to systemCmd
